refactor(UrlNode): split to ImageNode and ExternalLinkNode
