+++ src/block/node/index.ts
@@ -7,7 +7,8 @@ import { FormulaNodeParser } from './FormulaNode'
 import { DecorationNodeParser } from './DecorationNode'
 import { CodeNodeParser } from './CodeNode'
 import { BlankNodeParser } from './BlankNode'
-import { UrlNodeParser } from './UrlNode'
+import { ImageNodeParser } from './ImageNode'
+import { ExternalLinkNodeParser } from './ExternalLinkNode'
 import { GoogleMapNodeParser } from './GoogleMapNode'
 import { InternalLinkNodeParser } from './InternalLinkNode'
 import { IconNodeParser } from './IconNode'
@@ -23,12 +24,18 @@ import type { FormulaNode } from './FormulaNode'
 import type { DecorationNode } from './DecorationNode'
 import type { CodeNode } from './CodeNode'
 import type { BlankNode } from './BlankNode'
-import type { UrlNode } from './UrlNode'
+import type { ImageNode } from './ImageNode'
+import type { ExternalLinkNode } from './ExternalLinkNode'
 import type { GoogleMapNode } from './GoogleMapNode'
 import type { InternalLinkNode } from './InternalLinkNode'
 import type { IconNode } from './IconNode'
 import type { HashTagNode } from './HashTagNode'
 import type { PlainNode } from './PlainNode'
+
+/**
+ * @deprecated
+ */
+export type UrlNode = ImageNode | ExternalLinkNode
 
 export type LineNode =
   | QuoteNode
@@ -40,7 +47,8 @@ export type LineNode =
   | DecorationNode
   | CodeNode
   | BlankNode
-  | UrlNode
+  | ImageNode
+  | ExternalLinkNode
   | GoogleMapNode
   | InternalLinkNode
   | IconNode
@@ -79,7 +87,8 @@ export const convertToLineNodes = combineNodeParsers(
   StrongIconNodeParser,
   StrongNodeParser,
   DecorationNodeParser,
-  UrlNodeParser,
+  ImageNodeParser,
+  ExternalLinkNodeParser,
   IconNodeParser,
   GoogleMapNodeParser,
   InternalLinkNodeParser,
