+++ src/index.ts
@@ -5,7 +5,7 @@ export type { Title } from './block/Title'
 export type { CodeBlock } from './block/CodeBlock'
 export type { Table } from './block/Table'
 export type { Line } from './block/Line'
-export type { LineNode } from './block/node'
+export type { LineNode, UrlNode } from './block/node'
 export type { QuoteNode } from './block/node/QuoteNode'
 export type { HelpfeelNode } from './block/node/HelpfeelNode'
 export type { StrongImageNode } from './block/node/StrongImageNode'
@@ -15,7 +15,8 @@ export type { FormulaNode } from './block/node/FormulaNode'
 export type { DecorationNode, Decoration } from './block/node/DecorationNode'
 export type { CodeNode } from './block/node/CodeNode'
 export type { BlankNode } from './block/node/BlankNode'
-export type { ExternalLinkNode, ImageNode, UrlNode } from './block/node/UrlNode'
+export type { ImageNode } from './block/node/ImageNode'
+export type { ExternalLinkNode } from './block/node/ExternalLinkNode'
 export type { GoogleMapNode } from './block/node/GoogleMapNode'
 export type { InternalLinkNode } from './block/node/InternalLinkNode'
 export type { IconNode } from './block/node/IconNode'
